refactor: refactor services/agent_history
